requirements.txt
pandas
biopython
torch
transformers
datasets
esm
openmm
requests
biotite
pymupdf4llm
pymupdf
torch
torchvision
layoutparser
git+https://github.com/facebookresearch/detectron2.git@v0.5#egg=detectron2